+++ ios/Classes/SwiftStoreVersionCheckerPlugin.swift
@@ -1,0 +1,14 @@
+import Flutter
+import UIKit
+
+public class SwiftStoreVersionCheckerPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "store_version_checker", binaryMessenger: registrar.messenger())
+    let instance = SwiftStoreVersionCheckerPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    result("iOS " + UIDevice.current.systemVersion)
+  }
+}